Support multiple environment files on `forego run`
